Create common.js
Common.js contains snippets of code which are repeated over the database multiple times, and will likely be reused a bunch.
