+++ client/discord/commands/sector.js
@@ -38,62 +38,12 @@ exports.run = async (client, message, args, level) => { // eslint-disable-line n
         const title = `Sector`;
 
         if (!scanned) {
-            const sectorEmbed = client.extends.embed();
-            sectorEmbed.title = title;
-            sectorEmbed.description = `Cannot display sector until a \`scan\` has been completed.`;
-
-            const buttons = client.extends.row()
-                .addComponents(btnScan);
-
-            const sectorMessage = await message.channel.send({
-                embeds: [sectorEmbed],
-                components: [buttons]
-            });
-
-            const collector = client.extends.collector(sectorMessage, message.author);
-
-            collector.on('collect', async (i) => {
-                switch(i.customId) {
-                    case "btn_scan":
-                        await client.container.commands.get('scan').run(client, message, args, level);
-                        break;
-                }
-            });
+            await client.common.requireScan(message, title)
             return;
         }
 
         if (!sectorData.stellarObjects.length) {
-            const sectorEmbed = client.extends.embed();
-            sectorEmbed.title = title;
-            sectorEmbed.description = `There is nothing of interest in this sector, just empty space!
-
-You can \`jump\` to another sector or \`scan\` the sector again.`;
-
-            const buttons = client.extends.row()
-                .addComponents(btnScan)
-                .addComponents(btnJump)
-                .addComponents(btnMap);
-
-            const sectorMessage = await message.channel.send({
-                embeds: [sectorEmbed],
-                components: [buttons]
-            });
-
-            const collector = client.extends.collector(sectorMessage, message.author);
-
-            collector.on('collect', async (i) => {
-                switch(i.customId) {
-                    case "btn_scan":
-                        await client.container.commands.get('scan').run(client, message, args, level);
-                        break;
-                    case "btn_jump":
-                        await client.container.commands.get('jump').run(client, message, args, level);
-                        break;
-                    case "btn_map":
-                        await client.container.commands.get('map').run(client, message, args, level);
-                        break;
-                }
-            });
+            await client.common.emptySpace(message, title)
             return;
         }
